move functions, shorten loop, adjust docu
