+++ avaframe/ana5Utils/distanceTimeAnalysisCfg.ini
@@ -21,7 +21,8 @@ gateContours = 20
 # aperture angle in degrees
 aperture = 15
 
-# radar location field of view points min (x0, y0) and point giving direction (x1, y1) format: x0|x1|y0|y1
+# radar location field of view points min (x0, y0) and point giving direction (x1, y1)
+# format: x0|x1|y0|y1 needs to be in same projection as DEM
 radarLocation =
 
 #++++++++thalweg time diagram along avalanche path ++++++++++++
