+++ avaframe/tests/test_distanceTimeAnalysis.py
@@ -46,43 +46,6 @@ def test_setDEMOrigin():
     assert demOriginal['header']['cellsize'] == 2.
 
 
-def test_rotate():
-    """ test rotate a vector by an angle """
-
-    # setup required data
-    locationPoints = [[0., 0.], [0., 10.]]
-    theta = 45.
-    deg = True
-
-    # call function to be tested
-    rotatedLine = dtAna.rotate(locationPoints, theta, deg=deg)
-
-    sL = np.sqrt(50.)
-    assert rotatedLine[0][0] == 0.
-    assert -sL - 1.e-12 < rotatedLine[0][1] < -sL+1.e-12
-    assert rotatedLine[1][0] == 0.
-    assert sL - 1.e-12 < rotatedLine[1][1] < sL+1.e-12
-    assert isinstance(rotatedLine, list)
-    assert len(rotatedLine) == 2
-    assert len(rotatedLine[0]) == 2
-
-
-def test_cartToSpherical():
-    """ test converting to spherical coordinates """
-
-    # setup required inputs
-    X = 10.
-    Y = 10.
-    Z = np.sqrt(10.**2 + 10.**2)
-
-    # call function to be tested
-    r, phi, theta = dtAna.cartToSpherical(X, Y, Z)
-
-    assert r == np.sqrt(10.**2 + 10.**2 + Z**2)
-    assert phi == 45.
-    assert theta == 45.
-
-
 def test_radarMask(tmp_path):
     """ test creating radar mask array """
 
@@ -116,7 +79,7 @@ def test_avalancheMask():
     threshold = 4.19
 
     # call function to be tested
-    maskResType = dtAna.avalancheMask(resData, threshold)
+    maskResType = np.ma.masked_where(resData < threshold, resData)
 
     assert np.all(maskResType.mask[4,4:6]) == False
     assert np.all(maskResType.mask[5,4:6]) == False
@@ -138,31 +101,3 @@ def test_minRangeSimulation():
     losRange = dtAna.minRangeSimulation(flowF, rangeMasked, threshold)
 
     assert losRange == 4.
-
-
-def test_fetchFlowFields():
-    """ test fetching fields in a folder """
-
-    # setup required input
-    # get input data
-    dirPath = pathlib.Path(__file__).parents[0]
-    avaName = 'avaHockeyChannelPytest'
-    flowFieldsDir = dirPath / '..'/ '..' / 'benchmarks' / avaName / 'Outputs' / 'com1DFA' / 'peakFiles'
-    suffix = 'ppr'
-
-    # call function to be tested
-    flowFields = dtAna.fetchFlowFields(flowFieldsDir, suffix=suffix)
-    flowFields = sorted(flowFields)
-    print('flowFields', flowFields, sorted(flowFields))
-
-    assert flowFields[0].stem == 'release1HS_ent_dfa_d10bdc1e81_ppr'
-    assert flowFields[1].stem == 'release2HS_ent_dfa_e2145362b7_ppr'
-    assert len(flowFields) == 2
-
-    # call function to be tested
-    flowFields = dtAna.fetchFlowFields(flowFieldsDir)
-    flowFields = sorted(flowFields)
-    print('flowFields', flowFields, sorted(flowFields))
-
-    assert flowFields[0].stem == 'release1HS_ent_dfa_d10bdc1e81_pfd'
-    assert len(flowFields) == 6
